+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [unreleased]
+
+## [2.0.0] - 2020-12-03
 - Introduced `update(ByteBuffer)`.
 - Introduced `done(out: OutputStream, len: Int)`.
 - Introduced `done(out: ByteBuffer)`.
@@ -56,7 +58,8 @@ All notable changes to this project will be documented in this file.
 ## [1.0.0] - 2020-07-15
 - The first public release as dedicated project.
 
-[unreleased]: https://github.com/catap/scala-blake3/compare/v1.9.0...HEAD
+[unreleased]: https://github.com/catap/scala-blake3/compare/v2.0.0...HEAD
+[2.0.0]: https://github.com/catap/scala-blake3/compare/v1.9.0...v2.0.0
 [1.9.0]: https://github.com/catap/scala-blake3/compare/v1.8.1...v1.9.0
 [1.8.1]: https://github.com/catap/scala-blake3/compare/v1.8.0...v1.8.1
 [1.8.0]: https://github.com/catap/scala-blake3/compare/v1.7.1...v1.8.0